move TypeCompiler
