+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/TypeCompiler/BooleanTypeCompiler.java
@@ -29,54 +29,40 @@
  * and are licensed to you under the GNU Affero General Public License.
  */
 
-package com.splicemachine.db.impl.sql.compile;
-
-import java.sql.Types;
+package com.splicemachine.db.impl.sql.compile.TypeCompiler;
 
 import com.splicemachine.db.iapi.services.loader.ClassFactory;
+
 import com.splicemachine.db.iapi.sql.compile.CompilerContext;
-import com.splicemachine.db.iapi.types.TypeId;
 import com.splicemachine.db.iapi.types.DataTypeDescriptor;
+import com.splicemachine.db.iapi.types.TypeId;
+
 import com.splicemachine.db.iapi.sql.compile.TypeCompiler;
-import com.splicemachine.db.iapi.services.sanity.SanityManager;
+
 import com.splicemachine.db.iapi.reference.ClassName;
 
 /**
- * This class implements TypeCompiler for the SQL REF datatype.
+ * This class implements TypeCompiler for the SQL BOOLEAN datatype.
  *
  */
 
-public class RefTypeCompiler extends BaseTypeCompiler
+public class BooleanTypeCompiler extends BaseTypeCompiler
 {
-	/** @see TypeCompiler#getCorrespondingPrimitiveTypeName */
-	public String getCorrespondingPrimitiveTypeName()
-	{
-		if (SanityManager.DEBUG)
-			SanityManager.THROWASSERT("getCorrespondingPrimitiveTypeName not implemented for SQLRef");
-		return null;
-	}
-
 	/**
-	 * @see TypeCompiler#getCastToCharWidth
-	 */
-	public int getCastToCharWidth(DataTypeDescriptor dts, CompilerContext compilerContext)
-	{
-		if (SanityManager.DEBUG)
-			SanityManager.THROWASSERT( "getCastToCharWidth not implemented for SQLRef");
-		return 0;
-	}
-
-	/** @see TypeCompiler#convertible */
-	public boolean convertible(TypeId otherType, 
-							   boolean forDataTypeFunction)
-	{
-		return otherType.getJDBCTypeId() == Types.VARCHAR;
-	}
-
-	/**
-	 * Tell whether this type is compatible with the given type.
+	 * Tell whether this type (boolean) can be converted to the given type.
 	 *
-	 * @see TypeCompiler#compatible */
+	 * @see TypeCompiler#convertible
+	 */
+	public boolean convertible(TypeId otherType, boolean forDataTypeFunction)
+	{
+        return (otherType.isStringTypeId() || otherType.isBooleanTypeId());
+	}
+
+        /**
+         * Tell whether this type (boolean) is compatible with the given type.
+         *
+         * @param otherType     The TypeId of the other type.
+         */
 	public boolean compatible(TypeId otherType)
 	{
 		return convertible(otherType,false);
@@ -85,17 +71,57 @@ public class RefTypeCompiler extends BaseTypeCompiler
 	/** @see TypeCompiler#storable */
 	public boolean storable(TypeId otherType, ClassFactory cf)
 	{
-		return otherType.isRefTypeId();
+		/* Are the types the same or is other type a string */
+		if ( otherType.isBooleanTypeId() || otherType.isStringTypeId() )
+		{
+			return true;
+		}
+
+		/*
+		** If the other type is user-defined, use the java types to determine
+		** assignability.
+		*/
+		return userTypeStorable(getTypeId(), otherType, cf);
 	}
 
 	/** @see TypeCompiler#interfaceName */
 	public String interfaceName()
 	{
-		return ClassName.RefDataValue;
+		return ClassName.BooleanDataValue;
+	}
+
+	/**
+	 * @see TypeCompiler#getCorrespondingPrimitiveTypeName
+	 */
+
+	public String getCorrespondingPrimitiveTypeName()
+	{
+		/* Only numerics and booleans get mapped to Java primitives */
+		return "boolean";
+	}
+
+	/**
+	 * Get the method name for getting out the corresponding primitive
+	 * Java type.
+	 *
+	 * @return String		The method call name for getting the
+	 *						corresponding primitive Java type.
+	 */
+	public String getPrimitiveMethodName()
+	{
+		return "getBoolean";
+	}
+
+	/**
+	 * @see TypeCompiler#getCastToCharWidth
+	 */
+	public int getCastToCharWidth(DataTypeDescriptor dts, CompilerContext compilerContext)
+	{
+		return TypeCompiler.BOOLEAN_MAXWIDTH_AS_CHAR;
 	}
 
 	String nullMethodName()
 	{
-		return "getNullRef";
+		return "getNullBoolean";
 	}
 }
