+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/TypeCompiler/NumericTypeCompiler.java
@@ -29,7 +29,7 @@
  * and are licensed to you under the GNU Affero General Public License.
  */
 
-package com.splicemachine.db.impl.sql.compile;
+package com.splicemachine.db.impl.sql.compile.TypeCompiler;
 
 
 import com.splicemachine.db.iapi.sql.compile.CompilerContext;
