+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/TypeCompiler/TimeTypeCompiler.java
@@ -29,12 +29,9 @@
  * and are licensed to you under the GNU Affero General Public License.
  */
 
-package com.splicemachine.db.impl.sql.compile;
+package com.splicemachine.db.impl.sql.compile.TypeCompiler;
 
-import com.splicemachine.db.iapi.reference.SQLState;
 import com.splicemachine.db.iapi.services.loader.ClassFactory;
-
-import com.splicemachine.db.iapi.error.StandardException;
 
 import com.splicemachine.db.iapi.sql.compile.CompilerContext;
 import com.splicemachine.db.iapi.types.DataTypeDescriptor;
@@ -42,11 +39,10 @@ import com.splicemachine.db.iapi.types.TypeId;
 
 import com.splicemachine.db.iapi.sql.compile.TypeCompiler;
 
+import java.sql.Types;
 import com.splicemachine.db.iapi.reference.ClassName;
 
-import java.sql.Types;
-
-public class DateTypeCompiler extends BaseTypeCompiler
+public class TimeTypeCompiler extends BaseTypeCompiler
 {
 	/* TypeCompiler methods */
 	/**
@@ -56,26 +52,29 @@ public class DateTypeCompiler extends BaseTypeCompiler
 	 * date/time types; it may be generalized later for e.g.
 	 * comparison of any user type with one of its subtypes.
 	 *
-	 * @see TypeCompiler#convertible
+	 * @see TypeCompiler#convertible 
 	 */
-	public boolean convertible(TypeId otherType, boolean forDataTypeFunction) {
+	public boolean convertible(TypeId otherType,
+							   boolean forDataTypeFunction)
+	{
+
+		if (otherType.isStringTypeId() && 
+			(!otherType.isLOBTypeId()) &&
+			!otherType.isLongVarcharTypeId())
+		{
+			return true;
+		}
 
 
-        if (otherType.isStringTypeId() &&
-                (!otherType.isLongConcatableTypeId())) {
-            return true;
-        }
+		/*
+		** If same type, convert always ok.
+		*/
+		return (getStoredFormatIdFromTypeId() == 
+				otherType.getTypeFormatId());
+		   
+	}
 
-        // we can convert a date to a timestamp
-        return getTypeId().getJDBCTypeId() == Types.DATE && otherType.getJDBCTypeId() == Types.TIMESTAMP || (getStoredFormatIdFromTypeId() == otherType.getTypeFormatId());
-
-    }
-
-        /**
-         * Tell whether this type (date) is compatible with the given type.
-         *
-         * @param otherType     The TypeId of the other type.
-         */
+	/** @see TypeCompiler#compatible */
 	public boolean compatible(TypeId otherType)
 	{
 		return convertible(otherType,false);
@@ -96,7 +95,7 @@ public class DateTypeCompiler extends BaseTypeCompiler
 	public boolean storable(TypeId otherType, ClassFactory cf) {
         int otherJDBCTypeId = otherType.getJDBCTypeId();
 
-        return otherJDBCTypeId == Types.DATE || (otherJDBCTypeId == Types.CHAR) || (otherJDBCTypeId == Types.VARCHAR) || cf.getClassInspector().assignableTo(otherType.getCorrespondingJavaTypeName(), "java.sql.Date");
+        return otherJDBCTypeId == Types.TIME || (otherJDBCTypeId == Types.CHAR) || (otherJDBCTypeId == Types.VARCHAR) || (otherJDBCTypeId == Types.TIMESTAMP) || cf.getClassInspector().assignableTo(otherType.getCorrespondingJavaTypeName(), "java.sql.Time");
 
     }
 
@@ -112,7 +111,7 @@ public class DateTypeCompiler extends BaseTypeCompiler
 
 	public String getCorrespondingPrimitiveTypeName()
 	{
-		return "java.sql.Date";
+		return "java.sql.Time";
 	}
 
 	/**
@@ -120,28 +119,16 @@ public class DateTypeCompiler extends BaseTypeCompiler
 	 */
 	public int getCastToCharWidth(DataTypeDescriptor dts, CompilerContext compilerContext)
 	{
-		return 10;
+		return 8;
+	}
+
+	public double estimatedMemoryUsage(DataTypeDescriptor dtd)
+	{
+		return 12.0;
 	}
 
 	String nullMethodName()
 	{
-		return "getNullDate";
+		return "getNullTime";
 	}
-
-    @Override
-    public DataTypeDescriptor resolveArithmeticOperation(DataTypeDescriptor leftType, DataTypeDescriptor rightType, String operator) throws StandardException {
-        if (operator!= null && (operator.equals("*") || operator.equals("/"))) {
-            throw StandardException.newException(SQLState.LANG_DATE_TIME_MULT_DIV_PROHIBITED,
-                                                 TypeId.getBuiltInTypeId(Types.DATE).toParsableString(leftType));
-        }
-        DataTypeDescriptor returnType = leftType;
-        if (rightType != null && (rightType.getJDBCTypeId() == Types.DATE || rightType.getJDBCTypeId() == Types.TIMESTAMP)) {
-            if (operator!= null && operator.equals("+")) {
-                throw StandardException.newException(SQLState.LANG_DATE_TIME_ADDITION_PROHIBITED,
-                                                     TypeId.getBuiltInTypeId(Types.DATE).toParsableString(leftType));
-            }
-            returnType = DataTypeDescriptor.INTEGER;
-        }
-        return returnType;
-    }
 }
