+++ mem_sql/src/main/resources/com/splicemachine/db/modules.properties
@@ -61,7 +61,7 @@ derby.module.jdbcJ4=com.splicemachine.db.jdbc.Driver30
 derby.module.access.heap=com.splicemachine.derby.impl.store.access.hbase.HBaseConglomerateFactory
 derby.module.daemon=com.splicemachine.db.impl.services.daemon.SingleThreadDaemonFactory
 derby.env.jdk.cacheManagerJ6=6
-derby.module.tcf=com.splicemachine.db.impl.sql.compile.TypeCompilerFactoryImpl
+derby.module.tcf=com.splicemachine.db.impl.sql.compile.TypeCompiler.TypeCompilerFactoryImpl
 derby.env.classes.jdbcJ6=java.sql.Driver
 derby.env.classes.jdbcJ4=java.sql.Driver
 derby.env.jdk.cacheManagerJ1=1
